Replace http requests lib with local implementation
The original library didn't have a way to get the status of a request.
This was essential for some of the early challenges, and for generally
encouraging good practices. Unfortunately the original library seems to
not be maintained so I chose to reimplement the requests inside the
project. This adds the status codes but not much else.